0.36 work - 10 Feb
Played a testgame as Poland and fixed a number of things. Made an AI Bavaria more able to cope with not being a GP by the time the Brunswick congress has come and gone, tweaked some results of the Treaty of Szekes, tweaked some numbers on the Brunswick congress, added a decision for Saxony to be annexed to whatever German has it in its sphere. Further attempts to make the Russian war go  as planned - not yet successful entirely, as the Tatars often refuse to join the war.