0.36 work - 11 Feb
Tatars joining the war just vassalizes them, then the player has to call them in. Hopefully Russia does it automatically when it's AI. Made the cores disappear so that there wouldn't be random DON provinces turned over to RUS when the vassalization happens. No idea what it did that, but OK.