0.36 work - 8 April
Fixed and expanded the Mare Occidentis Nostrum decision - It is no longer repeatable. Added effects to make sure cores are added to all countries and their cores, added a tiny event chain for every relevant vassal to inherit them after a couple of days.
Also added decisions to accept the cultures found in the area.